requirements.txt
Flask>=2.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
google-cloud-storage>=2.0.0

